Fix relink process to detect acme certs.

+++ util/docker/web/startup_scripts/01_self_signed_ssl.sh
@@ -32,8 +32,13 @@ if [ ! -e "$ACME_DIR/ssl.crt" ]; then
 fi
 
 if [ ! -f "$ACME_DIR/ssl.crt" ]; then
-    ln -s "$ACME_DIR/default.key" "$ACME_DIR/ssl.key"
-    ln -s "$ACME_DIR/default.crt" "$ACME_DIR/ssl.crt"
+    if [ -f "$ACME_DIR/acme.crt" ]; then
+        ln -s "$ACME_DIR/acme.key" "$ACME_DIR/ssl.key"
+        ln -s "$ACME_DIR/acme.crt" "$ACME_DIR/ssl.crt"
+    else
+        ln -s "$ACME_DIR/default.key" "$ACME_DIR/ssl.key"
+        ln -s "$ACME_DIR/default.crt" "$ACME_DIR/ssl.crt"
+    fi
 fi
 
 chown -R azuracast:azuracast "$ACME_DIR" || true
